refactor: split out data-processing from dashboard.py
